full integrated VMD into flipbook
used pseudo-terminal utilities to run vmd since vmd as a subprocess auto terminates. This might make vmd less likely to work on windows systems but the user can still run the scripts manually.

Also changed how combined rmsx files were written to make them usable to vmd.

tweaked the visualization of the vmd grid files to make sure they lined up with the rmsx plots (not clamping at 5% and 95%.

I tried to make the minimum thickness larger and the scaling less but i think this is very usable as is and is more of a matter of taste.